# Conflicts:
#	web/src/locales/en.po
#	web/src/locales/pseudo-LOCALE.po

+++ authentik/outposts/controllers/docker.py
@@ -134,7 +134,8 @@ class DockerController(BaseController):
     def down(self):
         try:
             container, _ = self._get_container()
-            container.kill()
+            if container.status == "running":
+                container.kill()
             container.remove()
         except DockerException as exc:
             raise ControllerException from exc
